FEATURE: Update invites list after updating invite
This commit also extends the capabilities of the Invite model, which
is now being used in the create-invite modal.

The modal also uses a buffered proxy for most of the fields, except the
most complex ones which require some transformation between the server
and client serialization formats (i.e. topics and groups).